completed refactor of entry_type
